+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.30.3-0 / Unreleased
   * Replace jshint/jscs with eslint
   * Now testing on Puppeteer/Chromium instead of PhantonJS
+  * Deprecated loaderror and lazyload options now throw an error
 
 # 2.30.2 / 2019-01-13
   * Stop testing on IE 8 (no longer available on Saucelabs)